requirements.txt
Django==1.7.5
django-filter==0.9.2
django-nested-inline==0.3.4
djangorestframework==3.1.0
Markdown==2.6
Pillow==2.7.0
